test: HandlerAdapter에 대한 테스트 분리
